web_scraper/requirements.txt
requests
beautifulsoup4
scrapy
playwright
scrapy-playwright
pywebview
